pyproject.toml
[project]
name = "ai-mentor-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=0.6.3",
    "langchain>=0.3.19",
    "langchain-openai>=0.3.7",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "sentence-transformers>=3.4.1",
    "sqlmodel>=0.0.23",
    "typing-extensions>=4.12.2",
]
